api_yamdb/requirements.txt
django==2.2.16
djangorestframework==3.12.4
django-filter==21.1
djangorestframework-simplejwt==5.2.0
requests==2.26.0
PyJWT==2.1.0

asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
